Draft phase 4 implementation
