+++ meal_planner/ui/common.py
@@ -1,0 +1,3 @@
+from monsterui.all import TextT
+
+CSS_ERROR_CLASS = str(TextT.error)
